refactor: move LevelSet to its own module
